refactor: move components to dedicated directory
