+++ apps/main-web-app/src/routes/git/[...uri]/in-editor/_components/Sidebar.svelte
